Noogat Assignment/requirement.txt
python-pptx
google-generativeai
Pillow
python-dotenv
rich